Moded asdf update and plugin update
to the top of the script. Before the latest version numbers in asdf are checked need to make sure, that the latest plugins and asdf versions are installed, as these updates in some cases affect the result of **asdf latest**. Ruby 2.7.2 was not recognised as the latest ruby version with asdf 0.7.8 and the old plugin. Only after updating asdf and the ruby plugin was the latest ruby version correctly determined.